Refactor mem tests in multiple files
